add all-room-level, adjust-weapons, adjust-enemies
msh blm lengkap smua tp

+++ DungeonGunnerDav/Assets/Scripts/Dungeon/RoomEnemySpawnParameters.cs
@@ -7,7 +7,7 @@ public class RoomEnemySpawnParameters
 {
     public int maxTotalEnemiesToSpawn;
     public int maxConcurrentEnemies;
-    public int minSpawnInterval;
-    public int maxSpawnInterval;
+    public float minSpawnInterval;
+    public float maxSpawnInterval;
     public int wavesSurviveTimeInSeconds;
 }
